refactor: notification repository
